feature(logger-orchestrator): Refactors on management of severity.
